[python] Split user-facing Python code in bindings and interfaces
