+++ oif/interfaces/python/oif/interfaces/linear_solver.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from .core import OIFBackend, init_backend
+from oif.core import OIFBackend, init_backend
 
 
 class LinearSolver:
