--- src/oif/frontend_python/__init__.py
@@ -1,3 +1,0 @@
-"""
-Subpackage for the user-facing Python code
-"""
